Split ports.models into separate files
