+++ app/ports/models/port.py
@@ -1,11 +1,7 @@
-import urllib.request
-import ssl
 import json
-import datetime
 import os
 
 from django.db import models
-from django.contrib.postgres.fields import JSONField
 
 
 class Category(models.Model):
@@ -317,186 +313,3 @@ class Maintainer(models.Model):
             models.Index(fields=['github']),
             models.Index(fields=['name', 'domain'])
         ]
-
-
-class Builder(models.Model):
-    name = models.CharField(max_length=100, db_index=True)
-
-
-class BuildHistory(models.Model):
-    builder_name = models.ForeignKey(Builder, on_delete=models.CASCADE)
-    build_id = models.IntegerField()
-    status = models.CharField(max_length=50)
-    port_name = models.CharField(max_length=100)
-    time_start = models.DateTimeField()
-    time_elapsed = models.TimeField(null=True)
-    watcher_id = models.IntegerField()
-
-    class Meta:
-        indexes = [
-            models.Index(fields=['port_name', 'builder_name', '-build_id']),
-            models.Index(fields=['port_name', 'status', 'builder_name']),
-            models.Index(fields=['port_name', 'builder_name']),
-            models.Index(fields=['-time_start']),
-            models.Index(fields=['port_name']),
-            models.Index(fields=['status']),
-            models.Index(fields=['builder_name'])
-        ]
-
-    @classmethod
-    def populate(cls):
-        builders = Builder.objects.values_list('name', flat=True)
-        url_prefix = 'https://build.macports.org'
-
-        def get_url_json(builder_name, build_number):
-            return '{}/json/builders/ports-{}-builder/builds/{}'.format(url_prefix, builder_name, build_number)
-
-        def get_url_build(builder_name, build_number):
-            return '{}/builders/ports-{}-builder/builds/{}'.format(url_prefix, builder_name, build_number)
-
-        def get_data_from_url(url):
-            gcontext = ssl.SSLContext()
-            with urllib.request.urlopen(url, context=gcontext) as u:
-                data = json.loads(u.read().decode())
-            return data
-
-        def get_build_properties(array):
-            properties = {}
-            for prop in array['properties']:
-                properties[prop[0]] = prop[1]
-            return properties
-
-        def return_summary(builder_name, build_number, build_data):
-            data = {}
-
-            properties = get_build_properties(build_data)
-            port_name = properties['portname']
-            status = ' '.join(build_data['text'])
-            time_start = build_data['times'][0]
-            time_build = float(build_data['times'][1]) - float(build_data['times'][0])
-
-            data['name'] = port_name
-            data['url'] = get_url_build(builder_name, build_number)
-            data['watcher_id'] = properties['triggered_by'].split('/')[6]
-            data['watcher_url'] = properties['triggered_by']
-            data['status'] = status
-            data['builder'] = builder_name
-            data['buildnr'] = build_number
-            data['time_start'] = str(datetime.datetime.fromtimestamp(int(float(time_start)), tz=datetime.timezone.utc))
-            data['buildtime'] = str(
-                datetime.timedelta(seconds=int(float(time_build)))) if time_build is not -1 else None
-
-            return data
-
-        def load_database(data):
-            builder = Builder.objects.get(name=data['builder'])
-            build = BuildHistory()
-            build.port_name = data['name']
-            build.status = data['status']
-            build.build_id = data['buildnr']
-            build.time_start = data['time_start']
-            build.time_elapsed = data['buildtime']
-            build.builder_name = builder
-            build.build_url = data['url']
-            build.watcher_url = data['watcher_url']
-            build.watcher_id = data['watcher_id']
-            build.save()
-
-        for buildername in builders:
-            # fetch the last build first in order to figure out its number
-            last_build_data = get_data_from_url(get_url_json(buildername, -1))
-            last_build_number = last_build_data['number']
-            build_number_loaded = BuildHistory.objects.filter(builder_name__name=buildername).order_by('-build_id')
-            if build_number_loaded:
-                build_in_database = build_number_loaded[0].build_id + 1
-            else:
-                build_in_database = last_build_number - 9999
-
-            for build_number in range(build_in_database, last_build_number):
-                build_data = get_data_from_url(get_url_json(buildername, build_number))
-                build_data_summary = return_summary(buildername, build_number, build_data)
-                load_database(build_data_summary)
-
-
-# Contains the latest state of the JSON submitted by the mpstats for user
-class User(models.Model):
-    uuid = models.CharField(max_length=36, db_index=True)
-    osx_version = models.CharField(max_length=10)
-    macports_version = models.CharField(max_length=10)
-    xcode_version = models.CharField(max_length=10)
-    os_arch = models.CharField(max_length=10)
-    full_json = JSONField()
-    updated_at = models.DateTimeField(auto_now=True)
-
-
-class OSDistribution(models.Model):
-    osx_version = models.CharField(max_length=20, db_index=True)
-    month = models.IntegerField(db_index=True)
-    year = models.IntegerField(db_index=True)
-    users = models.ManyToManyField(User, related_name='users')
-
-
-class Commit(models.Model):
-    hash = models.CharField(max_length=50)
-    updated_at = models.DateTimeField(auto_now=True)
-
-
-class UUID(models.Model):
-    uuid = models.CharField(max_length=36, db_index=True)
-
-
-class Submission(models.Model):
-    user = models.ForeignKey(UUID, on_delete=models.CASCADE)
-    os_version = models.CharField(max_length=10)
-    xcode_version = models.CharField(max_length=10)
-    os_arch = models.CharField(max_length=20)
-    macports_version = models.CharField(max_length=10)
-    timestamp = models.DateTimeField()
-
-    class Meta:
-        indexes = [
-            models.Index(fields=['timestamp']),
-            models.Index(fields=['user']),
-            models.Index(fields=['user', '-timestamp']),
-            models.Index(fields=['user', 'timestamp']),
-            models.Index(fields=['os_version']),
-        ]
-
-    @classmethod
-    def populate(cls, json_object, timestamp):
-        uuid_obj, created = UUID.objects.get_or_create(uuid=json_object['id'])
-        sub = Submission()
-        sub.user = uuid_obj
-        sub.os_version = json_object['os']['osx_version']
-        sub.xcode_version = json_object['os']['xcode_version']
-        sub.os_arch = json_object['os']['os_arch']
-        sub.macports_version = json_object['os']['macports_version']
-        sub.timestamp = timestamp
-        sub.save()
-        return sub.id
-
-
-class PortInstallation(models.Model):
-    submission = models.ForeignKey(Submission, on_delete=models.CASCADE)
-    port = models.CharField(max_length=100)
-    version = models.CharField(max_length=100)
-    requested = models.BooleanField(default=False)
-
-    class Meta:
-        indexes = [
-            models.Index(fields=['submission']),
-            models.Index(fields=['port'])
-        ]
-
-    @classmethod
-    def populate(cls, port_json, submission_id):
-        ports = []
-        for port in port_json:
-            obj = PortInstallation()
-            obj.submission_id = submission_id
-            obj.port = port['name']
-            obj.version = port['version']
-            obj.requested = True if port.get('requested') == "true" else False
-            ports.append(obj)
-        PortInstallation.objects.bulk_create(ports)
-        return
